Update profile.py
Added all commands that should be necessary to set up the cluster in one go. Some commands are a bit redundant, but are set up to be explicit for all things to be done correctly. Also accounted for the machine list needing to be moved. Could be auto generated, but not testing that for now. Also made it so 12 compute nodes can be created.